separate ip functionalities into separate files
